requirements.txt
python-dotenv
google-genai
langchain
langchain-google-genai
pydantic